+++ src/env.d.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/triple-slash-reference */
 /// <reference path="../.astro/types.d.ts" />
 
 declare namespace App {
@@ -6,3 +7,5 @@ declare namespace App {
     runtime: 'cloudflare';
   }
 }
+
+export {};